test(spark): the block-key routing test needs a SESSION, not just pyspark
`spark_connect` caught it: `pyspark.sql.functions` asserts
`SparkContext._active_spark_context is not None`, so building a column
expression needs an active session rather than merely an importable pyspark.
`importorskip("pyspark")` passed locally only because the whole file skipped
there -- it failed in the one lane that actually has pyspark installed, which
is exactly the environment the guard was supposed to select for.

Moved next to the other JVM-routing tests, where the `spark` fixture provides a
session. The session-free unit file keeps the tests that genuinely need
nothing: slot dedup, exact-field slots, and the lambda arity contract.

The rest of the lane was green, including the four new ones that matter:
scoring parity at batch_size 1 / 3 / 10,000 and the end-to-end jar-only dedupe,
all exactly equal to the Python path.
